Don't use CMake's package config system
Given two CMake projects, A and B, where B depends on A:

When the export(PACKAGE <name>) command is used in project A's
CMakeLists.txt, CMake records the path to project A's build tree in a
registry entry (Windows) or ~/.cmake (Unix).

Project B's CMakeLists.txt can now use find_package(A) to search for an
AConfig.cmake file in project A's build tree. This would set various
variables, such as A_INCLUDE_DIRS and A_LIBRARIES, which allow B to use
A.

The advantage to this system is that project B does not have to know
anything else about project A--it uses find_package and some
standardized variable names, and that's it. The alternative is to use
add_subdirectory and magically know what variables to use.

The disadvantage is that multiple build trees can be recorded in the
registry. If they are distinguished by different version numbers, this
isn't a problem, but if they are incorrectly versioned (i.e., version
numbers not actually used yet, or someone forgot to bump the version
number), then the actual build tree that will be used by the dependent
project is undefined. This bit us in the ass, so we're removing it.