Prevent import libraries from being installed
On Win32, gcc generates a libsfp.dll.a import library. Initially I had
this installed along with the actual dll, but this file is only
necessary for gcc to know how to link with the dll. Once we've linked a
binary, the .dll.a does not need to be distributed with the binary. In
order to prevent higher level projects' installer packages from being
polluted with files that will never get used, I'm preventing archive
libraries from being installed by default.

If libsfp is ever used as a package in its own right, we may eventually
want to provide a facility for installing archives (i.e., static
libraries), particularly on linux. At that time, we can add logic which
checks the OS and the BUILD_SHARED_LIBS variable to decide which types
of targets to install.