Remove SFP_WRITE_ONE/SFP_WRITE_MULTIPLE
Previously, I allowed the user to select whether their I/O write function
handled individual octets or ranges of octets. This distinction was mis-
guided: individual octets is a special case of ranges of octets, making the
use of one function type adequate. Worse, I stored the function address in
a void*, which is undefined behavior, and caused problems when porting to
VS2017.